Pay attention to the order of symbol generation.
To efficiently compute LCAs later, we want to have the invariant that
nodes are always bound to nodes that have a lower id. Accordingly, we
need to always call gensym () for the parents first.

Note that some of the tests changed which variable names are output, due
to the ordering changes.