desugar let bindings to let bindings.
Otherwise we fail to generalize them!

This is still incomplete; we're still desugaring lets with patterns to
matches, which has the above problem -- we need to be slightly smarter
re: de-sugaring.

After de-sugaring we're getting an exception re: compare being applied to
functions. Isn't OCaml supposed to be type safe?