Get rid of hack with row binding edges.
This felt incredibly dubious at the time, and with this the failing
cases just cause a stack overflow.

I've stared at this enough that I'm now fairly certain that the root of
the problem is the same as the issue with recursive types. In both
cases, at some point a g-node is expanded where we would like it to
introduce a unification edge between one of the bottom nodes in the
structural interior and something else, but because it is in the
interior, no such link is made.