Revert change to row unification.
This was actually very, very broken.

Conceptually, it would be nice to define a normal form for rows where
they can just be unified naively, and then we deal with the equivalence
relation in one place. But my first pass had glaring problems and it's
not clear how to fix. We'll keep having a specialized unification
algorithm for rows, for now.