Allow leading commas in record syntax.
I'd been puzzling over this in the back of my mind for a while: I like
the regularity of allowing trailing commas in languages that do that,
like:

    foo = {
      bar,
      baz,
      quux,
    }

On a practical note, it makes the source easier to edit. But I've also
found that the comma-first notation is actually a good bit more
readable:

    foo =
      { bar
      , baz
      , quux
      }

In particular it makes it much easier to notice missing commas. Allowing
a leading comma gives us the best of both worlds:

    foo = {
      , bar
      , baz
      , quux
    }