Fix test failures.
The solution is to only bound the universally quantified variables
rigidly in the *parameter* of the coercion, and leave them flexible in
the result.

I'm still a little confused as to why we can get away with flattening
the structure like this (and half suspect the answer is we can't);
nothing here reflects the flip-floppy nature of higher-rank
types/contravariance beyond the first flip.