Do Quant normalization in the normalize module.
...rather than during unification.

The only reason it had previously been in unify is that it needs both
terms to know what to do -- so this adds a thing to normalize that looks
at both.

It would make a lot of sense to move the `Extend normalization here as
well, so unify can be even dumber.