Add a failing test.
This should work. I know why it's not: when deciding whether to rigid or
flexible bindings, we should actually be looking at whether we're in
positive or negative position in the type. If, like in this case, we see
an existential quantifier in a negative position, the two should cancel
and we should get a flexible binding. As it is, because the variable is
the root of the type, it just gets treated as rigid.