Update test.
I'd previously thought we had a bug here, but I realized that while
non-intuitive, this is actually correct. The type I expected was:

    (all a. a -> a -> a) -> ...

but we're getting:

    exist a. ((a -> a -> a) -> ...)

But these are actually equivalent! In the second case, the only way the
function can be called is if we have a function (a -> a -> a) that will
unify with *anything*, i.e. of type (all a. a -> a -> a).