Generate constraints for intrinsics more cleanly.
We just add the appropriate constraints to the environment, rather than
using the awful hack described in the (now deleted) comment.

This has the practical consequence that the constraint graphs are much
smaller, which results in both better performance (though it hasn't been
measured) and easier debugging of the constraint engine.