Update vim syntax re: identifiers.
for a while now we've allowed -, ?, and ! in identifiers, but the syntax
highlighting doesn't reflect it.

This at least gets the unquote operator doing the right thing, but I
need to look up how to get vim to recognize the differing set of
identifiers in general; right now it still highlights e.g. the "let"
in "let-foo", even though that should be one (non-keyword) identifier.