desugared ast: parametrize the metadata on exprs.
I want to delay converting the kinds back into AST form, but right now
we don't have a place to store intermediate kvars. The plan is to move
the type checker over to using `k_var Type.t` everywhere, but for that
we need to plumb the type parameter through to Expr.t.

There's no functional change yet; everything is still unit for now.