Fix the repl hang.
What was actually happening was that the globally declared get_line_io
was reading a line of text on startup. I don't really understand how
lwt's scheduling works -- apparently if you just evaluate this at top
level it will execute, but somehow it wasn't consuming the input except
when running the repl? Still not sure.

In any case, we now wrap IO in a function, so that we're guaranteed
side-effect free.