First attempt at separating grafting from merging.
Per the comments in graft_and_unify, the point of this is so that if
grafting succeeds but merging fails, then when we go back to do a
top-down merge, the grafting should still have made the subgraphs
isomorphic, which the top down merge doesn't try to handle.

The tests are failing with a segfault (!).

I don't know what's wrong, but I have a couple things to be suspicious
of:

* As written, if `copy_under` is passed a non-q node as the root, we
  could break our invariant that a `const node is necessarily inert.
* I have a gut feeling that this may make it possible to end up with a
  node bound on itself, though I've lost the reasoning.

It's also of course possible it's just something stupid.