Remove some stale comments.
These weren't quite correct, in that they didn't talk about the type
row.

Also, with the new Typebuilder module, the code is much more
declarative, and I no longer feel the commentary is really necessary --
before it was critical for understanding the code, because it was
otherwise so hard to follow.