Start some major internal refactoring.
I plan on moving over to a bidirectional type checker, for a couple
reasons:

* The solutions for the problems we're having with higher-kinded types
  are not obvious to me in the current context, but I think I know how
  they will work in a bidirectional setting.
* Right now error messages are garbage. A bidirectional checker will
  make fixing this much easier, to the point where it may offset having
  to re-do a bunch of work.

This patch itself doesn't get into the type checker yet. It starts off
by defining (the beginnings of) a new ast, "macro_ast." I plan for this
to replace desugared_ast for the purposes of type checking, maybe
entirely. As the name suggests, when we add a macro system, this is what
that macro system will see.