Some style updates
In particular, change semi-colon-first lists and records like:

    [ foo
    ; bar
    ; baz
    ]

into ones with trailing semicolons, like:

    [
      foo;
      bar;
      baz;
    ]

I've been doing this for newer code, since it allows easier editing due
to the fact that you can have a semicolon on the last element (Mule
allows commas before the first element for this reason). This commit
goes back and changes the old code to match.