Get rid of unused toplogical_sort dependency.
...since we've now written our own tsort module.

I had left this in place since I'd assumed we'd use it for #12
eventually, but we need to be able to handles, so it's unsuitable.

This also adds back the stdio dependency, which I guess is pulled in by
topological_sort? When I removed the latter uses of Stdio stopped
building.