Fix #16.
This cleans up the pretty printing for types so it does the right thing
again.

Along the way I discovered and fixed a bug, where if you wrote:

{} : {...exist r. r}

The type row would be existentially quantified, while if you left off
the type annotation it would be universally quantified. Now it is always
universally quantified unless actually specified.