Delete a bunch of code from util/
This deletes some dead code, but also reworks the way we're using maps
and sets. For most types we no longer bother to declare a FooMap, and
instead just use `(Foo.t, 'v, Foo.comparator_witness) Map.t` directly
instead. We use type aliases in a couple places.

There are a few cases where we can't quite do this. The main thing that
happens is if we have a map in part of a type definition that uses
[@@deriving sexp_of], we need the module wrappers so that the sexp_of_t
function for the map/set is available to ppx_deriving.