Better vim indentation.
Still not great, but ...better than the ocaml?

I think to do a good job we'd actually have to go find the matching
brace/etc, which I don't really want to do in vimscript right now.

At some point we'll have a pretty-printing tool that we can trigger on
save.