Find and add tests for some bugs.
Bug 1: we weren't handling the type annotation `rec a. a` at all. We
were still rejecting it (serendipitously), but we should report a proper
error here.

Bug 2: `rec a. all b. a` hangs the type checker, due to a mistake in the
way `subst` works; I caught this while reading the source. `copy` does a
separate substitution inside the quantifier, which means our `seen`
cache doesn't catch everything it should.

Both of these bugs now have tests, but neither are fixed.