Drive our expect tests with dune runtest.
This is nicer in a couple ways:

- It keeps going without `make -k`.
- It tells us *how many* tests failed.
- It opens the door to future improvements.

Note that we still wrap this in a call to `make check`, which is
actually important: it makes sure that main.exe is built before the
tests are actually run, since otherwise they'll all fail.

This also marks the currently failing tests as "xfail", which means
other test failures are not obscured by known ones.