A couple more enhancements to the graph rendering.
Specifically, render the graph at the beginning of each `check` and the
end of each `synth`.

There's also some unused code for traversing the context when rendering
the graph; per the comment, this ends up making a bunch of noise. But
I've left it there for now in the hopes of making use of it later.