Desugared ast: Factor the binding part of letrec out into its own type.
This is a preliminary step in dealing with an issue affecting the type
checking of records. Right now records are desugared into recursive let
bindings, with the body constructing the actual record. But we don't
actually want to lose track of the fact that it was a record literal.

So this factors out a piece of the letrec handling, so that it can be
shared by the record code path when we add that.