Start working on a parser-combinators library.
The commented out bits don't type check. Not sure whether there are
actual type errors or more bugs in the checker. Part of the goal of
writing this now is to flush out more checker bugs, and force me to
make the error messages good enough to actually work through this stuff.