Check: be smarter about pushing types under applications.
This gets the `eof` parser type checking; the problem previously was
that on hitting the application, it would try to synthesize the body,
which hits the known problems with inferring variants returned from
matches.

Now, it stays in checking mode when descending into the argument, so
this works as expected.