Kill a huge performance bug in the type checker.
Question: What the hell was I thinking when I wrote this?

Answer:

I probably thought we needed to keep IDs of bound variables unique,
but we actually don't, because we're checking for shadowing in
subst, and we never unify bound variables directly, only after
substituting. I knew I'd have to fix the perf issue eventually,
but put in the thing I was certain was correct, and it was fine while
working on small examples.

This made subst O(n^2) though, so it's completely untenable for
reasonable sized programs; importing the list module was taking
~3 seconds.