stdlib/list: factor out end-biased operations into sub modules.
Neat effect of this: the side of the list you want to work with is
a value that can be passed around. It also means fewer names for
things, and a conceptually smaller API surface.

It also gets us some neat things, like being able to do left.sort
vs. right.sort to vary ascending vs. descending.