Work on a perf issue with push_down_quants.
Right now this whole process is spectacularly inefficient, and many
of the stdlib modules just hang when trying to type check them. This
improves perf a bit, but we need more drastic speedups.

Specifically, because we're calling push_down_quants during
generalization, anything but the variables we're generalizing will
already be dealt with, so we don't need to do a recursive bottom-up
traverse of the tree each time. This makes a big difference on its own,
but the perf is still abysmal.

This patch also adds a big constant factor overhead in the case where
multiple variables are being generalized at the same time, since we'll
re-compute the contains map each time (though we may need to do so for
correctness anyway...)

We probably need to find a way to avoid rebuilding the contains map at
each generalization point alltogether; this is just not going to scale.