Correction & cleanup of const pattern checking logic.
The functional change here is that we do a `meet` to unify parameters,
rather than a `require_subtype`. This seems to not actually fix the test
that caught it, but here's why it's necessary:

If the parameter is not used in the body of the default case, then when
synth_leaf generalizes it, push_down_quants will push the quantifier
into the parameter part, so param will be `exist a. a`. *anything* will
check against this. That's actually as it should be, since the branch
placed no constraints on the parameter.

However, because the variable is quantified, any instantiations made
during the check will be discarded, and other branches will be free to
have *different* subtypes. This is not what we want. Therefore we want
to be doing a meet of all of the branches, rather than just checking
each of them against the default individually.