Relicense under Apache 2.0.
The more I think about it, the more I think LGPL is just going to
discourage people from using this. Languages have enough barriers to
adoption as-is.

Furthermore, I don't want to have to design the compiler around making
license compliance easy. I think static linking is generally a Good
Idea, and we'll probably want to be really aggressive with inlining,
which makes it hard to swap in alternate versions of a library -- which
means either a bunch of extra engineering effort on our part or
releasing source being the only realistic mechanism of compliance.

Finally, if we decide we *do* want a copyleft license of some sort, it
will be easier to add that later. By contrast, once people other than
me start contributing, I can't unilaterally re-license copylefted code
under a permissive license.