Switch to Francois Pottier's unionFind library.
...which hadn't been published when I started this project.

This doesn't really provide much benefit over the module we had *now*,
but I intend on moving back to MLF as a basis for the type system, and
following the unification algorithm more directly. That requires being
able to access the old graph, and this library supports pluggable
stores, whereas the module I wrote does not.

Note that the library doesn't have a direct equivalent of our `merge`
function, so we've retained that under the name `union_with`.