Further fussing wtih extract.
I think I'm going to try to split this in two phases: first, extract a
syntactic type with everything in quantifier bounds, and then separately
do the simplification. That should make this easier, and this patch
includes some work towards it. But my brain is fuzzy and am having
trouble sussing out all the details; this is a checkpoint, but it's not
otherwise a super great stopping point.