When merging two variables, always pick the minimum ID.
There is an invariant used by lca, that assumes that parent nodes have
lower ids than child nodes. There's a possible failure mode though,
where if two sibling nodes are merged, and we keep the ID of the larger
one, the children of the smaller node may also have IDs that are smaller
than the new parent. So we must pick the minimum ID in order to avoid
this.

Another case of "bug, but not the one I'm trying to debug now." Sigh.