Fix a bug in expand
Checking if a node in the constraint error was implemented incorrectly.
Specifically, we were collecting the parents on the way down, and
checking for set membership. But this doesn't work because it relies on
the assumption that all binders go to the *structural* root. But in failing
test cases, some go through g-nodes which are only *bound* on the root.
So being well-dominated is not quite enough for this approach to work.

Instead, we do a naive walk up the tree to check if we're under the G
node. We use Seen to memoize the results for efficiency.