Minor correction & refactor.
The downside of stubbing out to-be written functions with failwith is
that singular call sites will always type check -- as this one did, even
though we should have been extracting the `Lambda from deeper in the
tree. Factor out the code that does this earlier into a helper function
and do it here too.