Re-enable expect tests.
i.e. make them run as part of make check.

...and don't #SKIP any of them.

The SKIP was for the benefit of examples which weren't terminating,
which isn't something that comes up for the MLF type checker.
(Furthermore, we're not actually getting as far as type checking
for any of these...).

Right now the type checker still isn't rigged up in the right way
to actually be compatible with these again, but I want to do that
soon.